+++ packages/storage/index.js
@@ -2,6 +2,4 @@
 const storage = require('../../utils/storage')
 
 // export package
-module.exports = {
-	storage,
-}
+module.exports = storage